NodeMCU: turn on DHT22 only when used
Workaround: DHT22 may have various overheating and/or stability problems. Since
when connected to VCC it is constantly powered even when the ESP8266 is in deep
sleep mode, we now connect its VCC pin to GPIO and set it to high right before
reading the temperature. In accordance with the DHT22 datasheet we wait 1.1 s
after powering it up before reading it.

* DHT22 overheating and stability:
  http://www.kandrsmith.org/RJS/Misc/dht22_first_failure.html
* Unplug DHT22 to restore operativity:
  http://forum.arduino.cc/index.php?topic=355137.0
* DHT22 specs:
  http://www.electrodragon.com/w/DHT22_Digital_Humidity_and_Temperature_Sensor_(AM2302)